Move persistence and k8s
