+++ kotsadm/pkg/store/s3pg/airgap_store.go
@@ -6,7 +6,7 @@ import (
 	"github.com/pkg/errors"
 	"github.com/replicatedhq/kots/kotsadm/pkg/airgap/types"
 	airgaptypes "github.com/replicatedhq/kots/kotsadm/pkg/airgap/types"
-	"github.com/replicatedhq/kots/kotsadm/pkg/persistence"
+	"github.com/replicatedhq/kots/pkg/persistence"
 )
 
 func (s S3PGStore) GetPendingAirgapUploadApp() (*airgaptypes.PendingApp, error) {
