+++ kotsadm/pkg/store/s3pg/appstatus_store.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/replicatedhq/kots/kotsadm/pkg/appstatus"
-	"github.com/replicatedhq/kots/kotsadm/pkg/persistence"
 	appstatustypes "github.com/replicatedhq/kots/pkg/api/appstatus/types"
+	"github.com/replicatedhq/kots/pkg/persistence"
 )
 
 func (s S3PGStore) GetAppStatus(appID string) (*appstatustypes.AppStatus, error) {
