+++ kotsadm/pkg/store/s3pg/supportbundle_store.go
@@ -15,11 +15,11 @@ import (
 	"github.com/aws/aws-sdk-go/service/s3"
 	"github.com/aws/aws-sdk-go/service/s3/s3manager"
 	"github.com/pkg/errors"
-	"github.com/replicatedhq/kots/kotsadm/pkg/persistence"
 	kotss3 "github.com/replicatedhq/kots/kotsadm/pkg/s3"
 	"github.com/replicatedhq/kots/kotsadm/pkg/supportbundle/types"
 	supportbundletypes "github.com/replicatedhq/kots/kotsadm/pkg/supportbundle/types"
 	"github.com/replicatedhq/kots/pkg/logger"
+	"github.com/replicatedhq/kots/pkg/persistence"
 	troubleshootredact "github.com/replicatedhq/troubleshoot/pkg/redact"
 	"github.com/segmentio/ksuid"
 	"go.uber.org/zap"
