+++ kotsadm/pkg/store/s3pg/task_store.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/pkg/errors"
-	"github.com/replicatedhq/kots/kotsadm/pkg/persistence"
+	"github.com/replicatedhq/kots/pkg/persistence"
 )
 
 func (s S3PGStore) SetTaskStatus(id string, message string, status string) error {
